Fix #26: timeout requests to COPS.
This turns out to be easy with the 1.0 version of tokio: there is a built-in `timeout` primitive you can wrap around a request future before executing it.  So I have added a default timeout of 59 seconds, which is 1 second less than the NGL client times out at.

In debug builds, you can set the environment variable FRL_PROXY_TIMEOUT to your desired timeout (in milliseconds).  But in release builds the timeout is non-configurable.