Allow double slashes (//) in client URL paths.
Apparently the Adobe FRL apps, when they are talking directly to COPS, have a double slash in the front of their endpoint URLs.  To handle this, we need to use a patched version of the `warp` web framework that's tolerant of these double slashes.

This is worth a version bump to alpha.2.