Fix more cache bugs in schema usage.
- The source_addr was not being stored in the log session table.

This gets us to alpha.6 and so we update dependencies.

NOTE: A change was left out of the original version of this commit, but the tag was applied.  The commit has been amended and the tag re-applied.  If you pulled the old tag/commit, you will need to delete it and re-pull it to get the correct content.